+++ src/dynlib/steppers/__init__.py
@@ -6,7 +6,7 @@ from .registry import register, get_stepper, registry
 from .discrete import map
 from .ode import euler, rk4, rk45
 from .ode import ab2, ab3
-from .ode import bdf2_jit, bdf2
+from .ode import bdf2, bdf2_scipy
 
 __all__ = [
     "StepperCaps", "StepperMeta", "StepperInfo", "StepperSpec",